ml_project/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
pandas
scikit-learn
joblib